backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "A API central para o aplicativo de corrida fitness gamificado EvoRun."
authors = ["ViniciusBranco <viniciusbrancos@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.116.1"}
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.10"
pydantic = {extras = ["email"], version = "^2.11.7"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
python-multipart = "^0.0.20"
bcrypt = "^4.3.0"
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
